Don't override Meteor.user() to deal with guest logins, as many packages want to deal with the user object, even if the logged in user is a guest user.
The original intention for the override was to allow the
accounts-base "Sign-In" to appear. This commit still allows exactly
that, but does not return an empty user record when calling
Meteor.user() as guest. (Which was not the expected behaviour anyway,
I would argue. Because the method name implies that you get a user
record, regardless if the user is guest or not.)